move vmv fully into core
