+++ src/core/commit.rs
@@ -48,3 +48,43 @@ pub fn compute_polynomial_commitment<
     // Return `row_commitments` because they will come in handy for the opening proof
     (commitment, row_commitments)
 }
+
+/// Compute the size split (ν, σ) from variable count and chosen σ
+/// 2^ν is the number of rows; 2^σ the number of columns
+pub fn compute_nu(num_vars: usize, sigma: usize) -> usize {
+    if num_vars <= sigma * 2 {
+        sigma
+    } else {
+        num_vars - sigma
+    }
+}
+
+/// Compute the (Pedersen) commitments to the rows of the matrix M that is derived from coeffs `a`.
+/// This produces T` in the paper.
+pub fn commit_to_rows<E, M1, P>(
+  polynomial: &P,
+  sigma: usize,
+  nu: usize,
+  prover_setup: &ProverSetup<E>,
+) -> Vec<E::G1>
+where
+  E: Pairing,
+  M1: MultiScalarMul<E::G1>,
+  P: Polynomial<<E::G1 as Group>::Scalar, E::G1> + ?Sized,
+  E::G1: Group,
+  <E::G1 as Group>::Scalar: Field + Clone,
+{
+  // Use Γ₁[σ] as bases to commit each row of length 2^σ
+  debug_assert!(prover_setup.g1_vec().len() >= (1usize << sigma), "Γ1 length < 2^σ");
+  let bases = &prover_setup.g1_vec()[..1 << sigma];
+  let row_len = 1 << sigma;
+
+  let mut res = polynomial.commit_rows::<M1>(bases, row_len);
+
+  // Pad with identity elements if needed
+  while res.len() < (1 << nu) {
+      res.push(E::G1::identity());
+  }
+
+  res
+}
